Switch rank column from INTEGER to DECIMAL for fractional ranking
- Change rank column type from INTEGER to DECIMAL in both migrations
- Remove default value from create table migration
- Use manager.alter_table() for schema changes in rank migration
- Enables fractional ranking system for manual reordering (1.0, 1.5, 2.0, etc.)
- Supports unlimited precision like other numeric columns

+++ crates/migrations/src/m20231016_create_collection_to_entity.rs
@@ -129,7 +129,7 @@ impl MigrationTrait for Migration {
                     )
                     .col(
                         ColumnDef::new(CollectionToEntity::Rank)
-                            .integer()
+                            .decimal()
                             .not_null(),
                     )
                     .foreign_key(
